% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdb_local_tools.R
\name{approximate_lookup}
\alias{approximate_lookup}
\title{Approximate Lookup Function}
\usage{
approximate_lookup(
  df,
  column,
  search_string,
  scorer1 = "lv",
  scorer2 = "lcs",
  limit = 10
)
}
\arguments{
\item{df}{A data frame containing the data to search.}

\item{column}{A string specifying the name of the column to search.}

\item{search_string}{A string to search for in the specified column.}

\item{scorer1}{A string specifying the method to use for the initial scoring (default is "lv" for Levenshtein distance).}

\item{scorer2}{A string specifying the method to use for the secondary scoring (default is "lcs" for Longest common substring distance).}

\item{limit}{An integer specifying the maximum number of results to return (default is 10).}
}
\value{
A data frame containing the best matches, with columns for the original name, the best matching synonym, the index of the match, and the best score.
}
\description{
This function performs an approximate string matching lookup on a specified column of a data frame.
}
\examples{
\dontrun{
df <- data.frame(name = c("Name1", "Name2"), 
                 shift1.ppm. = c(1.0, 2.0), 
                 accession = c("A1", "A2"), 
                 synonyms = c("Syn1, Syn1Alt", "Syn2, Syn2Alt"))
approximate_lookup(df, "name", "name1")
}
}
