% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdb_local_tools.R
\docType{data}
\name{hmdb_nmr_1h}
\alias{hmdb_nmr_1h}
\title{hmdb_nmr_1h
A dataframe containing the assignment of 1H NMR spectra related to metabolites}
\format{
An object of class \code{data.frame} with 3208 rows and 13 columns.
}
\usage{
hmdb_nmr_1h
}
\description{
hmdb_nmr_1h
A dataframe containing the assignment of 1H NMR spectra related to metabolites
}
\keyword{datasets}
