% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdb_local_tools.R
\name{multiplet_query}
\alias{multiplet_query}
\title{Multiplet Query Function}
\usage{
multiplet_query(
  query,
  nmrdb = hmdb.local::hmdb_nmr_1h,
  metabolites = hmdb.local::urine_metabolites
)
}
\arguments{
\item{query}{A list of queries, where each query is a vector of the form \code{c(from, to, type)}.}

\item{nmrdb}{A data frame containing the NMR data with columns \code{shift1.ppm.}, \code{accession}, and \code{type}.}

\item{metabolites}{A data frame containing the metabolites data with columns \code{accession} and \code{name}.}
}
\value{
A data frame containing the matching metabolites, with columns for \code{accession}, \code{name}, and \code{similarity}, sorted by similarity in descending order.
}
\description{
This function performs range queries on NMR data and returns a data frame with results ranked by similarity.
}
\examples{
\dontrun{
query <- list(c(1.0, 1.5, "type1"), c(2.0, 2.5, "type2"))
nmrdb <- data.frame(shift1.ppm. = c(1.1, 2.1), accession = c("A1", "A2"), type = c("type1", "type2"))
metabolites <- data.frame(accession = c("A1", "A2"), name = c("Met1", "Met2"))
multiple_query(query, nmrdb, metabolites)
}
}
